more shims for torchaudio 0.7.2 vs 0.8.0

+++ audiotools/core/audio_signal.py
@@ -173,9 +173,7 @@ class AudioSignal(
         return self.to(device)
 
     def write(self, audio_path, batch_idx=0):
-        torchaudio.save(
-            audio_path, self.audio_data[batch_idx], self.sample_rate, precision=32
-        )
+        torchaudio.save(audio_path, self.audio_data[batch_idx], self.sample_rate)
         return self
 
     def deepcopy(self):
